+++ src/views/UserSearch/UserSearch.js
@@ -65,7 +65,7 @@ class UserSearch extends React.Component {
 
   static defaultProps = {
     idPrefix: 'users-',
-    visibleColumns: ['status', 'name', 'barcode', 'patronGroup', 'username', 'email'],
+    visibleColumns: ['active', 'name', 'barcode', 'patronGroup', 'username', 'email'],
   };
 
   constructor(props) {
@@ -273,7 +273,7 @@ class UserSearch extends React.Component {
     }
 
     const resultsFormatter = {
-      status: user => (
+      active: user => (
         <AppIcon app="users" size="small" className={user.active ? undefined : css.inactiveAppIcon}>
           {user.active ? <FormattedMessage id="ui-users.active" /> : <FormattedMessage id="ui-users.inactive" />}
         </AppIcon>
@@ -379,7 +379,7 @@ class UserSearch extends React.Component {
                           contentData={users}
                           totalCount={count}
                           columnMapping={{
-                            status: intl.formatMessage({ id: 'ui-users.active' }),
+                            active: intl.formatMessage({ id: 'ui-users.active' }),
                             name: intl.formatMessage({ id: 'ui-users.information.name' }),
                             barcode: intl.formatMessage({ id: 'ui-users.information.barcode' }),
                             patronGroup: intl.formatMessage({ id: 'ui-users.information.patronGroup' }),